React to 401 unauthorized correctly
When using HTTP basic authentication in Jolokia and you provide the wrong credentials from the client object, you get back a 401 in the headers and an empty body. The request method assumes a body hash will always be present, which it's not if you get a 401 response.

This distinguishes between HTTP level status codes and Jolokia status codes so the user can still check RemoteError and inspect the result.